# docs/swagger/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the FeatherWings daemon.",
        "title": "FeatherWings API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/download/backup": {
            "get": {
                "security": [
                    {
                        "ServerJWT": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Downloads"
                ],
                "summary": "Download server backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Signed backup token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/download/file": {
            "get": {
                "security": [
                    {
                        "ServerJWT": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Downloads"
                ],
                "summary": "Download server file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Signed file token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "router.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "NodeToken": {
            "description": "Supply the node's bearer token from `config.yml` using the `Authorization: Bearer \u003ctoken\u003e` header.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerJWT": {
            "description": "Signed JWTs issued by the Panel for server-scoped operations (uploads, downloads, websockets). Pass in the `token` query parameter.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}